fix processor tests

+++ process/process_series.py
@@ -62,10 +62,9 @@ def docker_runtime(task: Task, folder: str) -> bool:
 
         return option_dict
 
-
     real_folder = Path(folder)
 
-    if config.get_runner() == "docker": 
+    if config.get_runner() == "docker":
         # We want to bind the correct path into the processor, but if we're inside docker we need to use the host path
         # TODO: don't hardcode this!! can we use the mercure_data volume that docker knows about instead?
         # we don't want to just mount the whole mercure_data in though.
@@ -84,7 +83,7 @@ def docker_runtime(task: Task, folder: str) -> bool:
         return False
     additional_volumes: Dict[str, Dict[str, str]] = decode_task("additional_volumes")
     environment = decode_task("environment")
-    environment = {**environment, **dict(MERCURE_IN_DIR="/data",MERCURE_OUT_DIR="/output")}
+    environment = {**environment, **dict(MERCURE_IN_DIR="/data", MERCURE_OUT_DIR="/output")}
     arguments = decode_task("arguments")
 
     # Merge the two dictionaries
@@ -173,7 +172,7 @@ def process_series(folder) -> None:
         if config.get_runner() == "nomad" or config.mercure.process_runner == "nomad":
             # Use nomad if we're being run inside nomad, or we're configured to use nomad regardless
             processing_success = nomad_runtime(task, folder)
-        elif config.get_runner() in ("docker","systemd"):
+        elif config.get_runner() in ("docker", "systemd"):
             # Use docker if we're being run inside docker or just by systemd
             processing_success = docker_runtime(task, folder)
         else:
@@ -183,7 +182,7 @@ def process_series(folder) -> None:
         logger.error("Processing error.")
         logger.error(traceback.format_exc())
     finally:
-        if config.get_runner() == "docker":
+        if config.get_runner() in ("docker", "systemd") and config.mercure.process_runner != "nomad":
             logger.debug("Docker processing: immediately move results")
             move_results(folder, lock, processing_success, needs_dispatching)
             shutil.rmtree(folder, ignore_errors=True)
